Move/update SSL verification management command
